Optimize cv detection by using multithreading

+++ backend/tools/Exec.py
@@ -29,87 +29,81 @@ def UpdateDict(Dict1, Dict2):
 
 
 Result = {}
-def videoAnalyser(
-    video_path: str,
+def save_images(
+    cap: cv2.VideoCapture,
+    ModelDir: str,
     bChkH: bool,
     bChkBW: bool,
     bChkSplit_then_BW: bool,
     bChkNobarSplit_then_BW: bool,
     bChkBlackback: bool,
-    output_folder: str,
-    ModelDir: str
+    output_folder, subdir,
+    stopEvent: threading.Event
 ):
     global Result
 
-    print(Fore.GREEN, 'analysis_video', Style.RESET_ALL)
+    model_cls = YOLO(Path(ModelDir).joinpath('models_cls_videoHua.pt').as_posix())
+    model_clsBw = YOLO(Path(ModelDir).joinpath('modelm-cls_screen_w_rec_basic.pt').as_posix())
+    model_detect_splitScreen = YOLO(Path(ModelDir).joinpath("model_splitScreen.pt").as_posix()) #[2024-8-6]model_splitScreen.pt
 
-    print(Fore.GREEN, 'analysis_video', video_path, bChkH, bChkBW, bChkSplit_then_BW, Style.RESET_ALL)
+    # Initialize frame counter
+    frame_count = 0
+    # Start capturing frames
+    while not stopEvent.is_set():
+        ret, frame = cap.read()
+        if not ret:
+            break  # Break the loop if there are no frames left to read
 
-    # 展开得到frame
-    subdir = os.path.basename(video_path).split('.')[0]
-    extract_frames(video_path, output_folder, subdir)
-
-    # 检查花屏
-    lst_file = os.listdir(Path(output_folder).joinpath(subdir).as_posix())
-    # lst_mapH = {0:'hua',1:'off',2:'onGood'}
-    lst_outputH:List[str] = []
-    if bChkH:
-        model_clsHua = YOLO(Path(ModelDir).joinpath('models_cls_videoHua.pt').as_posix())
-        for file in lst_file:
-            result_class = model_clsHua(Path(output_folder).joinpath(subdir, file).as_posix(), verbose=False)[0]
-            # print(file, result_class.probs.top1)
+        # 检查花屏
+        lst_outputH:List[str] = []
+        if bChkH:
+            # 分析当前帧
+            result_class = model_cls(frame, verbose=False)[0]
+            # 如果检测到异常（花屏），保存图像
             if result_class.probs.top1 == 0:
+                file = save_image(frame_count, frame, output_folder, subdir)
                 lst_outputH.append(file)
-            else:
-                pass
-        UpdateDict(
-            Dict1 = Result,
-            Dict2 = {'lst_outputH': lst_outputH}
-        )
+            UpdateDict(
+                Dict1 = Result,
+                Dict2 = {'lst_outputH': lst_outputH}
+            )
 
-    # 检查黑白
-    # lst_mapBW = {0:'black',1:'good',2:'white'}
-    lst_outputB:List[str] = []
-    lst_outputW:List[str] = []
-    if bChkBW:
-        model_clsBw = YOLO(Path(ModelDir).joinpath('modelm-cls_screen_w_rec_basic.pt').as_posix())
-        for file in lst_file:
-            result_class = model_clsBw(Path(output_folder).joinpath(subdir, file).as_posix(), verbose=False)[0]
+        # 检查黑白
+        lst_outputB:List[str] = []
+        lst_outputW:List[str] = []
+        if bChkBW:
+            # 分析当前帧
+            result_class = model_clsBw(frame, verbose=False)[0]
+            # 如果检测到异常（黑白），保存图像
             if result_class.probs.top1 == 0:
+                file = save_image(frame_count, frame, output_folder, subdir)
                 lst_outputB.append(file)
             elif result_class.probs.top1 == 2:
                 # 由于特殊图像造成模型的判断问题,这里先用蒙特卡罗判断一下是否是白色
-                isBlack = is_mostly_black(Path(output_folder).joinpath(subdir, file))
+                isBlack = is_mostly_black(ret)
                 if isBlack:
+                    file = save_image(frame_count, frame, output_folder, subdir)
                     lst_outputB.append(file)
                 else:
+                    file = save_image(frame_count, frame, output_folder, subdir)
                     lst_outputW.append(file)
             else:
                 pass
-        UpdateDict(
-            Dict1 = Result,
-            Dict2 = {'lst_outputB': lst_outputB, 'lst_outputW': lst_outputW}
-        )
+            UpdateDict(
+                Dict1 = Result,
+                Dict2 = {'lst_outputB': lst_outputB, 'lst_outputW': lst_outputW}
+            )
 
-    # 分屏+检查黑白
-    lst_outputSplitB:List[str] = []
-    # lst_outputSplitW:List[str] = [] # 暂不考虑
-    if bChkSplit_then_BW:
-        # 这里需要对frame进行分割，然后对分割后的图像进行白色判断
-        model_detect_splitScreen = YOLO(Path(ModelDir).joinpath("model_splitScreen.pt").as_posix()) #[2024-8-6]model_splitScreen.pt
-        for file in lst_file:
-            path_pic = Path(output_folder).joinpath(subdir, file).as_posix()
-            dir_splitTmp = Path(CurrentDir).joinpath('tmp/clsW_split_train_pred/test_output').as_posix()
-            #
-            shutil.rmtree(dir_splitTmp, ignore_errors=True)
-            os.makedirs(dir_splitTmp, exist_ok=True)
-
-            img = cv2.imread(path_pic)
-            height, width = img.shape[:2]
+        # 分屏+检查黑白
+        lst_outputSplitB:List[str] = []
+        # lst_outputSplitW:List[str] = [] # 暂不考虑
+        if bChkSplit_then_BW:
+            # 这里需要对frame进行分割，然后对分割后的图像进行白色判断
+            height, width = frame.shape[:2]
+            # 分析当前帧
+            result_class = model_detect_splitScreen(frame, verbose=False)[0]
             # print(f'height: {height}, width: {width}')
-            #
-            results = model_detect_splitScreen(path_pic, verbose=False)[0] #, save=True
-            for n, r in enumerate(results):
+            for n, r in enumerate(result_class):
                 # print(r.boxes)  # print the Boxes object containing the detection bounding boxes
                 for i, box in enumerate(r.boxes):
                     if box is not None:
@@ -119,59 +113,105 @@ def videoAnalyser(
                         # [2024-2-22]增加一段,对类型的判断
                         x,y,w,h = n_cords
                         # print(Fore.MAGENTA, f'box{i}: {n_cords}, conf: {conf}', Style.RESET_ALL)
-                        # 注意下面LR和UD的判断,是分开写的.但如果保存成一样的文件名,其实可以合并
+                        # 注意下面LR和UD的判断,是分开写的.
                         if w < h:
                             # 竖向,从上到下分割
-                            L_img = img[0:height, 0:int(x)]
-                            R_img = img[0:height, int(x+w):width]
-                            cv2.imwrite(Path(dir_splitTmp).joinpath(f'det_r{n}_{i}_L.jpg').as_posix(),L_img)
-                            cv2.imwrite(Path(dir_splitTmp).joinpath(f'det_r{n}_{i}_R.jpg').as_posix(),R_img)
-                            isBlack = is_mostly_black(Path(dir_splitTmp).joinpath(f'det_r{n}_{i}_L.jpg').as_posix()) | is_mostly_black(Path(dir_splitTmp).joinpath(f'det_r{n}_{i}_R.jpg').as_posix())
+                            L_img = frame[0:height, 0:int(x)]
+                            R_img = frame[0:height, int(x+w):width]
+                            isBlack = is_mostly_black(L_img) | is_mostly_black(R_img)
                             if isBlack:
+                                file = save_image(frame_count, frame, output_folder, subdir)
                                 lst_outputSplitB.append(file)
                         else:
                             # 横向,从左到右分割,这时y和h是有用的
-                            # cropped_img = img[int(y-h/2):int(y+h/2), int(x-w/2):int(x+w/2)]
-                            #
-                            upper_img = img[0:int(y), 0:width]
-                            downer_img = img[int(y):height, 0:width]
-                            # U_img = img[0:height, 0:x]
-                            # D_img = img[0:height, x+w:width]
-                            cv2.imwrite(Path(dir_splitTmp).joinpath(f'det_r{n}_{i}_upper.jpg').as_posix(), upper_img)
-                            cv2.imwrite(Path(dir_splitTmp).joinpath(f'det_r{n}_{i}_downer.jpg').as_posix(), downer_img)
-                            isBlack = is_mostly_black(Path(dir_splitTmp).joinpath(f'det_r{n}_{i}_upper.jpg').as_posix()) | is_mostly_black(Path(dir_splitTmp).joinpath(f'det_r{n}_{i}_downer.jpg').as_posix())
+                            # cropped_img = frame[int(y-h/2):int(y+h/2), int(x-w/2):int(x+w/2)]
+                            upper_img = frame[0:int(y), 0:width]
+                            downer_img = frame[int(y):height, 0:width]
+                            isBlack = is_mostly_black(upper_img) | is_mostly_black(downer_img)
                             if isBlack:
+                                file = save_image(frame_count, frame, output_folder, subdir)
                                 lst_outputSplitB.append(file)
-        UpdateDict(
-            Dict1 = Result,
-            Dict2 = {'lst_outputSplitB': lst_outputSplitB}
-        )
+            UpdateDict(
+                Dict1 = Result,
+                Dict2 = {'lst_outputSplitB': lst_outputSplitB}
+            )
 
-    # [2024-8-15]无bar分屏+检查黑白
-    lst_outputNobarSplitB:List[str] = []
-    if bChkNobarSplit_then_BW:
-        for file in lst_file:
-            if nobar_split_half_black(Path(output_folder).joinpath(subdir, file).as_posix()):
+        # [2024-8-15]无bar分屏+检查黑白
+        lst_outputNobarSplitB:List[str] = []
+        if bChkNobarSplit_then_BW:
+            if nobar_split_half_black(frame):
+                file = save_image(frame_count, frame, output_folder, subdir)
                 lst_outputNobarSplitB.append(file)
-        UpdateDict(
-            Dict1 = Result,
-            Dict2 = {'lst_outputNobarSplitB': lst_outputNobarSplitB}
-        )
+            UpdateDict(
+                Dict1 = Result,
+                Dict2 = {'lst_outputNobarSplitB': lst_outputNobarSplitB}
+            )
 
-    # [2024-8-15]桌面来电底色
-    lst_outputBlackback:List[str] = []
-    if bChkBlackback:
-        for file in lst_file:
-            if is_mostly_black(Path(output_folder).joinpath(subdir, file).as_posix()):
+        # [2024-8-15]桌面来电底色
+        lst_outputBlackback:List[str] = []
+        if bChkBlackback:
+            if is_mostly_black(frame):
+                file = save_image(frame_count, frame, output_folder, subdir)
                 lst_outputBlackback.append(file)
-        UpdateDict(
-            Dict1 = Result,
-            Dict2 = {'lst_outputBlackback':lst_outputBlackback}
+            UpdateDict(
+                Dict1 = Result,
+                Dict2 = {'lst_outputBlackback':lst_outputBlackback}
+            )
+
+        frame_count += 1
+        print(f'frame_count: {frame_count}')
+    print('Result:', Result)
+
+
+def videoAnalyser(
+    video_path: str,
+    bChkH: bool,
+    bChkBW: bool,
+    bChkSplit_then_BW: bool,
+    bChkNobarSplit_then_BW: bool,
+    bChkBlackback: bool,
+    output_folder: str,
+    ModelDir: str,
+    stopEvent: threading.Event
+):
+    global Result
+
+    print(Fore.GREEN, 'analysis_video', Style.RESET_ALL)
+    print(Fore.GREEN, 'analysis_video', video_path, bChkH, bChkBW, bChkSplit_then_BW, Style.RESET_ALL)
+
+    # 设置输出文件夹
+    subdir = Path(video_path).parent.stem
+    #os.makedirs(os.path.join(output_folder, subdir), exist_ok=True)
+
+    # 加载视频流
+    cap = cv2.VideoCapture(0 if video_path is None else video_path)
+
+    # 启动图像处理线程
+    thread = threading.Thread(
+        target=save_images,
+        args=(
+            cap,
+            ModelDir,
+            bChkH,
+            bChkBW,
+            bChkSplit_then_BW,
+            bChkNobarSplit_then_BW,
+            bChkBlackback,
+            output_folder,
+            subdir,
+            stopEvent
         )
+    )
+    thread.start()
+    thread.join()
+
+    # 释放视频捕获对象
+    cap.release()
 
 
 adbRecord = None
-StopEvent = None
+StopRecordEvent = threading.Event()
+StopAllEvent = threading.Event()
 def RecordAndPull(
     SavePath_AD: str,
     SaveDir_PC: str,
@@ -185,10 +225,11 @@ def RecordAndPull(
     RecPeriod: int = 180,
 ):
     global adbRecord
-    global StopEvent
+    global StopRecordEvent
+    global StopAllEvent
     i = 0
     analysingThreads = []
-    while not StopEvent.is_set():
+    while not (StopRecordEvent.is_set() or StopAllEvent.is_set()) :
         try:
             adbRecord = Record(SavePath_AD, RecPeriod)
             adbRecord.wait()
@@ -197,24 +238,28 @@ def RecordAndPull(
             i += 1
             OldName = Path(SaveDir_PC).joinpath(Path(SavePath_AD).name).as_posix()
             NewName = Path(SaveDir_PC).joinpath(f"{i}{Path(SavePath_AD).suffix}").as_posix()
-            if Path(NewName).exists():
-                os.remove(NewName)
-            os.rename(OldName, NewName)
+            if Path(OldName).exists():
+                if Path(NewName).exists():
+                    os.remove(NewName)
+                os.rename(OldName, NewName)
         except Exception as e:
             print(f"RecordAndPull error: {e}")
         finally:
+            if not Path(NewName).exists():
+                continue
             analysingThread = threading.Thread(
                 target = videoAnalyser,
-                args = (NewName, bChkH, bChkBW, bChkSplit_then_BW, bChkNobarSplit_then_BW, bChkBlackback, output_folder, ModelDir)
+                args = (NewName, bChkH, bChkBW, bChkSplit_then_BW, bChkNobarSplit_then_BW, bChkBlackback, output_folder, ModelDir, StopAllEvent)
             )
             analysingThreads.append(analysingThread)
             analysingThread.start()
     else:
-        for analysingThread in analysingThreads:
-            try:
-                analysingThread.join()
-            except:
-                pass
+        if not StopAllEvent.is_set():
+            for analysingThread in analysingThreads:
+                try:
+                    analysingThread.join()
+                except:
+                    pass
 
 
 def Exec(
@@ -230,7 +275,8 @@ def Exec(
     RecPeriod: int = 180,
 ):
     global adbRecord
-    global StopEvent
+    global StopRecordEvent
+    global StopAllEvent
     global Result
 
     # Set the save location
@@ -240,9 +286,6 @@ def Exec(
     # Reboot server
     adbReboot = Reboot()
     adbReboot.wait()
-
-    # Event to signal the recording thread to stop
-    StopEvent = threading.Event()
 
     # Start the screen recording thread
     recordingThread = threading.Thread(
@@ -256,10 +299,15 @@ def Exec(
         TaskCMD,
         shell = True
     )
-    adbTask.wait()
+    # Wait for the task to finish or for the stop event to be set
+    while not StopAllEvent.is_set():
+        if adbTask.poll() is not None:
+            break
+    else:
+        adbTask.terminate()
 
     # Signal the recording thread to stop and wait for it to finish
-    StopEvent.set()
+    StopRecordEvent.set()
     if adbRecord is not None:
         adbRecord.terminate()
     recordingThread.join()
